// src/Components/CatalogoWeb/PuntoVenta/Model/Variant.d.ts
export type Variant = {
    Id?: number;
    Barcode?: string | null;
    Product_Id?: number;
    Description: string;
    Color?: string | null;
    Price?: number | null;
    PromotionPrice?: number | null;
    CostPerItem?: number | null;
    Stock?: number | null;
    ExpDate?: string | null;
    MinStock?: number | null;
    OptStock?: number | null;
};
 